Phase 4 partial: Extract shaders and useFractalRenderer hook
